Make the language switcher a toggle button
This duplicates the language switcher on the tugraz.at site.

Compared to the previous implementation we now show the next language
and use the translations for the next language, so that users only knowing
english can understand it.

In case of more languages we just cycle through them, which isn't ideal but
good enough for now.