Use the rollup resolve plugin for finding package asssets
Instead of hardcoding the paths of the node_modules directory we use the node-resolve
plugin of rollup to find the root of the corresponding JS package and calcucate a path from
there.

Resolving a package requires calling an async function, so we have to use await in the rollup config.
Luckily rollup supports configs wrapped in a promise, so we just have to wrap it in a function which returns
a promise.