Load the keycloak-js libary from the keycloak server instead of bundling it
This makes sure the library version works with the server version.

Ideally we'd also want to check the version so we can error out if a too
old keycloak server is on the other side, but I don't see any public API changes
that would allow us to know if the library is too old.

Fixes #32