Move the QR detection out of the redraw loop
We now queue it and in the callback check if the scan is still active,
and go from there. This means we redraw all the time and detect mostly
in the background.

This results in worse perf with the old library since that isn't async, but
we can always revert if needed.