Stop depending on puppeteer
While it's nice to have a fixed chrome version it slows down the yarn install,
downloads a lot of things, and the chrome version depends on various
system packages so one needss to have chromium installed anyway.

This means developers that want to work on tests need firefox and chdomium installed.

Sadly I couldn't find a way to make browsers optional in case the developer
only has one installed.