Refine spacing between elements in responsive mode; Make sticky menu more touch friendly
Now that we have clearer separated sections in responsive mode the different
spacing stand out so unify them a bit.

Also in case the menu is in sticky mode it's kinda small and hard to hit via touch
so add some padding there and remove some margin from the headline instead to keep
the overall height the same.