LoginButton: Don't show a logged-in/logged-out status during login
In case the login process takes some time we would show a "login"
button which makes the user try to click it. There is a good chance
that aroudn that time we get logged in, so the click will log the user
out.

Avoid that by not showing anything during logging in.