getPackagePath: handle the current package not being installed
In case we want to get the package path of the current package, but it
isn't in node_modules (as is the case with yarn2) we need to find it
ourselves.

In case the package.json in the current dir has the same name we just use
it as a base. This is after rollup lookup failed, so this is just a fallback.