language-select: add a small demo for how to override translations at runtime
Uses a separate namespace that gets used over the default one, so we can override,
but also remove the overrides again when needed.

This has the problem that all clones have the same store, so this overrides
the translations for all clones and not just the one providing the overrides.