i18n: set more fallback languages
We use i18n.languages in the app-shell to see which languages are available
but this only contains a list of languages it will search in (means, currently
set language, default one and all fallbacks).

We can just set all languages we know as last fallback, so i18n.languages
always contains all languages.

A bit hacky...