translation overrides: Use the real tag name when looking for overrides
Instead of using the namespace which is hardcoded we use the tag name that was
used to register the element (or the "real" tag name in case of a scoped element).

This way if AppShell is regsitered as "dbp-some-app" it can still be overriden by
specifying "dbp-some-app" in the overrides object.