slot templates: fix templates being created in the wrong shadow root
This makes problems with the new scoped elements polyfill, which now correctly
makes globally registered elements not visible in the shadow roots. Instead
we have to add the tempaltes at the level of the templates so they get their
global web components and only then transplant them into the light dom of the
nested web components with their own registries.