app-shell: allow trailing paths in the router
In case the path ending with the activity ID has more trailing path elements, we don't fall
back to the default route but just ignore them (and save them in the router state).

This allows passing in extra information via the path without breaking the routing,
and in case we want to forward the routing to the activities in the future we can use
the extra path elements we store for that.

For this to work with our current logic we have to stop comparing path strings and compare
the resulting computed state everywhere instead.