auth: move back to the keycloak package from npm again
Instead of fetching it from the configured keycloak server.

We had this initially, but switched to the server variant to make
sure we are always compatible with the keycloak version.

Now upstream keycloak has deprecated this strategy and in Keycloak 22
the server provided js bundle will be removed.

So switch back to the npm package.