app-shell: rework document title content a bit
Show the activity name first, and append the application name. This makes
the document title start with the active activity which might be more usefull
to the user. Especially if the user has multiple activities open and wants
to select the right one.

Special case the welcome activity by not show that activity name. It's not very helpful
as it's the same for all applications, so just skip it and only show the topic/app name.

Use the short version for the application name. There isn't much space usually where the
document title is shown, so if there is a short version, use it.