data-table-view: update override CSS for header arrows and expand buttons
The internals of datatables had changed which made the override CSS no longer
match properly and lead to arrows being repeated all over the place.

This just removes the CSS styling for the sorting arrows, since there are too
many things/states to override and this will likely fail again with future updates.
And the default arrows look good enough now.

For the exand buttons adjust the CSS selectors to match again, since that seems
simple enough. And because the default is green/red, which doesn't match our CI at all.