keycloak: fix a potential race condition when logging in
newer keycloak fails when an instance is inited multiple times,
and with the async login it could happend that two paths would
call _ensureInit() and because of the loading would both call
init().

Fix this by waiting on the same promise in all cases.

This also removes the workaround for https://gitlab.tugraz.at/dbp/topics/library/issues/41
since that would no longer with newer keycloak not allowing init() to be called
two times. Assume this is fixed in either Keycloak or Chrome.