person-select: remove json-ld initialization
We now build the API URLs manually. The initialization code is sadly depending
on the timing of the json-ld fetch, so try to hack around the select2 init code
to get a somewhat similar result.

Ideally this would track all the state that is needed for select2 and just re-init
it when the state changes, so we could check for updates on every property update
like auth, or url.