person-select: take buildUrlData() override into account when fetching one resource
In case the person-select gets subclassed and extra query parameters injected for fetching
additional data then we also need to do this when we try to restore the state through
fetching a single resource by ID. So we can re-use the same function pass a dummy search term
for now.

This fixes the pre-selection in case the display format function requires extra data
requested via buildUrlData() (for example via localData).