Move to 0.x.y for versioning for all packages
Otherwise the patch versions are not considered compatible
according to semver.

This can easily lead to confusion, in the context of this mono repo
where packages are pulled from npm instead of locally because they
don't match the requirements. Or when all dependants need to update
their requirements on each release.