Add a new @dbp-toolkit/dev-utils package
private for now, so only accessible extern if used as
a submodule.

This contains various helpers for rollup builds. Compare to the current
aproach we can add proper dependencies and apps don't need to make sure
that all dependencies are in their package.json.

We might extend it in the the future to contain linter, formatter,
typechecker configs.