Drop @dbp-toolkit/typescript-example
This was meant to evaluate using Typescript for the implementation and
to have a working build/bundle setup for typescript users as a starting point.

We have never switched to typescript though and also have no plans to switch
to it. And typescript never had any problems using the toolkit here besides updates
for package upgrades, so I assume there isn't anything needed from our side if we
continue publishing normal esm JS.

Remove to get rid of one more build config/variant we have to care about.