Get rid of "disabled-layout" again
Instead there is the "disable-layouts" attribute instead.
In case we want to add more layouts in the future the one attribute
for disabling one layout can't be expanded easily and applications
will usually need layouts nor none at all. So if we would add a third
one this would auto enable the layout switched for apps that have it
disabled.

Avoid that by using an attribue that disables the layout switcher
all together.