CI: build each package in isolation
This uncovers missing dependency problems and dependency cycles,
since npm hoists all depenencies in the workspace, so a dependency
might get installed via another package in the workspace without
us noticing that it is missing.

And only when the released pacakge is used in isolation will the build fail.

Work around by installing the depenencies one by one and running a build.
Sadly a bit slow, but hopefully prevents such errors in the future.

Removes various cyclic dependencies while at it (and might break some demos,
but nothing we can easily do about it atm)