app-shell: stop pushing lang/entry-point-url to activities
We've replaced this with a "pull" model via subscribe long ago.
The code tried to disable it in case there is a subscription, but
the format is not correct (the default has no mapping and no ":"),
so there would be lots of warnings due to setting an attribute that
is subscribed.

In case this breaks something, make sure to set

subscribe="lang,entry-point-url"

to your activity.