app-shell: provide a base-url for activities
This absolute URL under which the activity is being loaded.
This can be used by activities to pass as a callback URLs for APIs,
so that after passing control to an external website the website
can return to the activity.

More specifically this will be used by mono to set a return URL
for the payment provider.