common: fix getAssetURL() in case it is bundled into the entry point
We currently assumed it was always in the chunk directory, but that
is no always the case. I don't really know a good way to decide
if it is called from a chunk or not, so assume we are if there
is a "/shared/" parent path element.

This breaks if the entry point is itself under /shared, but that
should be unlikely.

Also add some more testing.