CI: fail if running i18next-parser results in updates
In case there are translation keys in the .js files that are
not in the translation .json files, or the otherway around,
or if the translation files are not sorted, then this step
will fail.

This should avoid the translation files getting out of sync
with the source code.

To fix it simply run "npm run i18next" at the toplevel of the repo.