keycloak: reset the requested login status after login/logout
In case a login is triggered and we redirect, but the user doesn't finish
the login and goes back, then the browser restores the state of the page
as it was.

Since the requestedLoginStatus is now logged-in, changing it again wont notify
the keycloak auth component. Make sure to reset the status once received,
so we get notified on changes again.