person-select: allow clearing the selection
up until now once a person was selected there was no UI way to clear things, except via the clear()
js API method. In some cases where a person is optional (like in a form) it is needed
that the user can unselect a person again.

We couldn't think of a reason why this can't be available always, so enable it always.