LangMixin: allow specifying an alternative property name
In case we have multiple translations in a class hierarchy we
need to register the different translations with different properties
or they will override each other.

This allows one to pass an alternative property name for the i18next
instance in a subclass.

Of course there are still problems when you add translations in a base
class at a later point etc, but I don't know how to best solve this.