form-elements: in case the required validator triggers then stop
In case a required field has a custom validator and it is empty we
would potentially show both errors if it was empty. We think that
in most cases this is more confusing as helpful in the empty case,
so in case a required input is empty we just show the required error
and nothing else.

Once the input has a value we validate more.