pdf-viewer: provide function to import/setup pdf.js
For pdf.js to work we need to copy and configure the worker.
To avoid every dependency having to do that, do it once in
the pdf-viewer and allow others to reuse it.

While some apps don't need the viewer part, they can just use
the importPdfJs() function alone if wanted.o