dev-utils: introduce system to automatically copy package assets on build
Many of our packages require that certain files they ship, or that some of their
dependencies ship, be at a certain path in the final build (for example icons,
worker files, images, css, etc).

To avoid every reverse dep of those packages having to copy the logic to
put the files in the right place, add some metadata to each package.json
which contains the source and destination of all those files.

The package to be built can then recursively collect the metadata from
all its runtime dependencies and pass the information to the rollup
copy plugin.

This means if, for example, a package deep in the dependency tree starts
requiring new files, we only have to change it in the package and not
any of the reverse dependencies.