rollup: add metadata to package.json for setting up the url plugin
We currently use the url plugin to copy files into the shared directory
with has in the filename, and to get a path to that file at runtime.

Compared to the normal copied files these get cached better and are
also shared between builds.

This adds metadata to the package.json, similar to assets, which defines
which files, when imported, should get copied to the shared dir.

In addition getUrlOptions() in dev-utils collects the metadata of all
dependencies and provides merged options to the url plugin.