tabulator: downloadExcel: remove truncation of filenames/worksheet names
It truncated the filename and worksheet name, but the real reason was that
xlsx doesn't allow more than 31 chars for worksheet names and filenames are fine.

Now that we use exceljs, error handling of too long work sheet name is slightly
different: it will just emit a warning and truncate the worksheet name internally.

Just remove the truncation on our end and let exceljs handle things.