Port from i18next-parser to i18next-cli
i18next-parser is deprecated and users should move to
i18next-cli.

Port the config to match the old one + adjust the npm commands
to do the same thing. The error message when translations have
changed is a bit confusing (when running "npm run i18next"),
but we can't change that.