dev-utils: add support for packages which don't have a default export
In that case resolve() doesn't give us any path and just fails.
Add another fallback which searches for the package in node_modules
manually.

An alternative would be to:

* in case we just look for metadata in package.json we can ignore
  those packages
* in case we have a path for a file we can use that as an export
  when calling resolve().

But for now this is good enough.

Motivated by instantsearch.css v8.9.0 which added explicit exports
and has no default one.